Review: add ifdefs, make docs explicit, cleanup docstrings, rename files
refs #29632
