+++ framework/doc/content/source/mfem/materials/MFEMGenericFunctorVectorMaterial.md
@@ -9,10 +9,10 @@
 ## Overview
 
 `MFEMGenericFunctorVectorMaterial` defines one or more vector material properties with values
-obtained from coefficients on one or more subdomains of the mesh, given by the `blocks` parameter
+obtained from coefficients on one or more subdomains of the mesh, given by the [!param](/Materials/MFEMGenericFunctorVectorMaterial/blocks) parameter
 if provided, or applied to the entire mesh if missing. The vector material properties are named
-according to members in the `prop_names` parameter, with respective coefficients used to get property
-values given by the members of `prop_values`. The coefficients in `prop_names` must be vector-valued.
+according to members in the [!param](/Materials/MFEMGenericFunctorVectorMaterial/prop_names) parameter, with respective coefficients used to get property
+values given by the members of [!param](/Materials/MFEMGenericFunctorVectorMaterial/prop_values). The coefficients in [!param](/Materials/MFEMGenericFunctorVectorMaterial/prop_names) must be vector-valued.
 
 !syntax parameters /Materials/MFEMGenericFunctorVectorMaterial
